Add polymorphic serialization test for `InputElement`.
